+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 - delete dry_run config (use dry_run option).
 - change layout config.
+- add usage in help.
 
 ## [0.3.1] - 2022-04-05