+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## [Unreleased]
 
+## [0.4.0] - 2022-04-09
+
 - delete dry_run config (use dry_run option).
 - change layout config.
 - add usage in help.
@@ -31,7 +33,8 @@
 
 - initial release (build, server, info, list, output, md, njk, relurl).
 
-[Unreleased]: https://github.com/ankys/homura_deno/compare/v0.3.1...HEAD
+[Unreleased]: https://github.com/ankys/homura_deno/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/ankys/homura_deno/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/ankys/homura_deno/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/ankys/homura_deno/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/ankys/homura_deno/compare/v0.1.0...v0.2.0